Updated library versions, fixed mojo test not running
