+++ src/test/java/core/ApplicationMojoTest.java
@@ -1,53 +1,38 @@
 package core;
 
-import org.apache.maven.plugin.Mojo;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugin.testing.AbstractMojoTestCase;
-import org.apache.maven.plugin.testing.ConfigurationException;
-import org.codehaus.plexus.configuration.PlexusConfiguration;
-import org.codehaus.plexus.util.ReaderFactory;
-import org.codehaus.plexus.util.xml.Xpp3Dom;
-import org.codehaus.plexus.util.xml.Xpp3DomBuilder;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import junit.framework.TestFailure;
+import junit.framework.TestResult;
+import junit.framework.TestSuite;
+import junit.textui.TestRunner;
+import org.junit.jupiter.api.Test;
 
-import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.io.OutputStream;
+import java.io.PrintStream;
+import java.util.Collections;
+import java.util.Enumeration;
+import java.util.List;
+import java.util.stream.Collectors;
 
-public class ApplicationMojoTest extends AbstractMojoTestCase {
-    public void testMojoGoal() throws Exception {
-        Exception e = runMojo(lookupMojo("start", Paths.get("src", "test", "resources", "pom2", "pom.xml")));
-        assertNotNull(e);
-        assertEquals(e.getCause().getMessage(), "project.uri is not properly defined in config.properties");
-    }
+import static org.assertj.core.api.Assertions.assertThat;
 
-    public Mojo lookupMojo(String goal, Path configPath) throws Exception {
-        PlexusConfiguration configuration = null;
-        if (configPath != null) {
-            try {
-                Xpp3Dom thisProject = Xpp3DomBuilder.build(ReaderFactory.newXmlReader(new File("pom.xml")));
-                configuration = extractPluginConfiguration(thisProject.getChild("artifactId").getValue(), Xpp3DomBuilder.build(ReaderFactory.newXmlReader(new File("src/test/resources/pom2/pom.xml"))));
-            } catch (ConfigurationException ignored) {
+public class ApplicationMojoTest {
+    private static final PrintStream NULL_STREAM = new PrintStream(new OutputStream() {
+        @Override
+        public void write(int b) {
 
-            }
         }
-        return lookupMojo(goal, configuration);
+    });
+
+    @Test
+    void testApplicationMojo() {
+        TestResult result = new TestRunner(NULL_STREAM).doRun(new TestSuite(ApplicationMojoTestInternal.class));
+        assertNoErrors(result.errors());
+        assertNoErrors(result.failures());
     }
 
-    public Mojo lookupMojo(String goal, PlexusConfiguration configuration) throws Exception {
-        Xpp3Dom thisProject = Xpp3DomBuilder.build(ReaderFactory.newXmlReader(new File("pom.xml")));
-        return lookupMojo(thisProject.getChild("groupId").getValue(), thisProject.getChild("artifactId").getValue(), thisProject.getChild("version").getValue(), goal, configuration);
-    }
-
-    @Nullable
-    public Exception runMojo(@NotNull Mojo mojo) {
-        try {
-            mojo.execute();
-            return null;
-        } catch (MojoExecutionException | MojoFailureException ex) {
-            return ex;
-        }
+    @SuppressWarnings("RedundantOperationOnEmptyContainer" /* inspection is wrong */)
+    private void assertNoErrors(Enumeration<TestFailure> errors) {
+        List<TestFailure> list = Collections.list(errors);
+        assertThat(list).overridingErrorMessage(list.stream().map(TestFailure::trace).collect(Collectors.joining("\n\n"))).isEmpty();
     }
 }
